\name{gplot.snatm}
\alias{gplot.snatm}
\encoding{UTF-8}
\title{\code{gplot} function of \pkg{sna} version 1.5.}
\description{\code{gplot} function of \pkg{sna} version 1.5.}
\usage{gplot.snatm(dat, g = 1, gmode = "digraph", diag = FALSE, label =
                 c(1:dim(dat)[2]), coord = NULL, jitter = TRUE, thresh
                 = 0, usearrows = TRUE, mode = "fruchtermanreingold",
                 displayisolates = TRUE, interactive = FALSE, xlab =
                 NULL, ylab = NULL, xlim = NULL, ylim = NULL, pad =
                 0.2, label.pad = 0.5, displaylabels = !missing(label),
                 boxed.labels = TRUE, label.pos = 0, label.bg =
                 "white", vertex.sides = 8, vertex.rot = 0,
                 arrowhead.cex = 1, label.cex = 1, loop.cex = 1,
                 vertex.cex = 1, edge.col = 1, label.col = 1,
                 vertex.col = 2, label.border = 1, vertex.border = 1,
                 edge.lty = 1, label.lty = NULL, vertex.lty = 1,
                 edge.lwd = 0, label.lwd = par("lwd"), edge.len = 0.5,
                 edge.curve = 0.1, edge.steps = 50, loop.steps = 20,
                 object.scale = 0.01, uselen = FALSE, usecurve = FALSE,
                 suppress.axes = TRUE, vertices.last = TRUE, new =
                 TRUE, layout.par = NULL, ...)}
\details{See ?gplot(sna).}
\author{Carter T. Butts}