\name{help.content.closeness}
\alias{help.content.closeness}
\encoding{UTF-8}
\title{Closeness Centrality of \proglang{R}-help Content-Based Communication Network}
\description{Closeness centrality of \proglang{R}-help content-based communication network calculated by Pajek
(\url{http://vlado.fmf.uni-lj.si/pub/networks/pajek/})}
\usage{data(help.content.closeness)}
\format{A vector of closeness centralities.}
\author{Angela Bohn \email{angela.bohn at gmail.com}}
\examples{
data(help.content.closeness)
}