\name{help.subjects.closeness}
\alias{help.subjects.closeness}
\encoding{UTF-8}
\title{Closeness Centrality of R-help Subject-Based Communication Network}
\description{Closeness centrality of R-help subject-based communication network calculated by Pajek
(\url{http://vlado.fmf.uni-lj.si/pub/networks/pajek/})}
\usage{data(help.subjects.closeness)}
\format{A vector of closeness centralities.}
\author{Angela Bohn \email{angela.bohn at gmail.com}}
\examples{
data(help.subjects.closeness)
}