\name{not.take.memory}
\alias{not.take.memory}
\encoding{UTF-8}
\title{Negative Alias Memory}
\description{A database of aliases found by \code{find.aliases()} considered to be incorrect.}
\usage{data(not.take.memory)}
\format{A list where each list element contains a vector of negative aliases.}
\author{Angela Bohn \email{angela.bohn at gmail.com}}
\examples{
data(not.take.memory)
}