\name{reduce.clusters}
\alias{reduce.clusters}
\title{Reduce number of clusters}
\description{Reduce number of clusters by sorting out those matched by first name or last name only.}
\usage{reduce.clusters(clusters)}
\arguments{
\item{clusters}{A list containing the alias clusters (result of \code{findclusters()}).}
}
\value{A list containing the remaining clusters.}
\author{Angela Bohn \email{angela.bohn at gmail.com}}
