\name{take.memory}
\alias{take.memory}
\encoding{UTF-8}
\title{Positive Alias Memory}
\description{A database of aliases found by \code{find.aliases()} considered to be correct.}
\usage{data(take.memory)}
\format{A list where each list element contains a vector of positive aliases.}
\author{Angela Bohn \email{angela.bohn at gmail.com}}
\examples{
data(take.memory)
}