\name{devel.subjects.closeness}
\alias{devel.subjects.closeness}
\title{Closeness Centrality of R-devel Subject-Based Communication Network}
\description{Closeness centrality of R-devel subject-based communication network calculated by Pajek
(\url{http://vlado.fmf.uni-lj.si/pub/networks/pajek/})}
\usage{data(devel.subjects.closeness)}
\format{A vector of closeness centralities.}
\author{Angela Bohn \email{angela.bohn at gmail.com}}
\examples{
data(devel.subjects.closeness)
}