\name{gplot.snatm}
\alias{gplot.snatm}
\title{\code{gplot} function of \pkg{sna} version 1.5.}
\description{\code{gplot} function of \pkg{sna} version 1.5.}
\usage{gplot.snatm(dat, g = 1, gmode = "digraph", diag = FALSE, label =
                 c(1:dim(dat)[2]), coord = NULL, jitter = TRUE, thresh
                 = 0, usearrows = TRUE, mode = "fruchtermanreingold",
                 displayisolates = TRUE, interactive = FALSE, xlab =
                 NULL, ylab = NULL, xlim = NULL, ylim = NULL, pad =
                 0.2, label.pad = 0.5, displaylabels = !missing(label),
                 boxed.labels = TRUE, label.pos = 0, label.bg =
                 "white", vertex.sides = 8, vertex.rot = 0,
                 arrowhead.cex = 1, label.cex = 1, loop.cex = 1,
                 vertex.cex = 1, edge.col = 1, label.col = 1,
                 vertex.col = 2, label.border = 1, vertex.border = 1,
                 edge.lty = 1, label.lty = NULL, vertex.lty = 1,
                 edge.lwd = 0, label.lwd = par("lwd"), edge.len = 0.5,
                 edge.curve = 0.1, edge.steps = 50, loop.steps = 20,
                 object.scale = 0.01, uselen = FALSE, usecurve = FALSE,
                 suppress.axes = TRUE, vertices.last = TRUE, new =
                 TRUE, layout.par = NULL)}
\arguments{
\item{dat}{See ?gplot(sna)}
\item{g}{See ?gplot(sna)}
\item{gmode}{See ?gplot(sna)}
\item{diag}{See ?gplot(sna)}
\item{label}{See ?gplot(sna)}
\item{coord}{See ?gplot(sna)}
\item{jitter}{See ?gplot(sna)}
\item{thresh}{See ?gplot(sna)}
\item{usearrows}{See ?gplot(sna)}
\item{mode}{See ?gplot(sna)}
\item{displayisolates}{See ?gplot(sna)}
\item{interactive}{See ?gplot(sna)}
\item{xlab}{See ?gplot(sna)}
\item{ylab}{See ?gplot(sna)}
\item{xlim}{See ?gplot(sna)}
\item{ylim}{See ?gplot(sna)}
\item{pad}{See ?gplot(sna)}
\item{label.pad}{See ?gplot(sna)}
\item{displaylabels}{See ?gplot(sna)}
\item{boxed.labels}{See ?gplot(sna)}
\item{label.pos}{See ?gplot(sna)}
\item{label.bg}{See ?gplot(sna)}
\item{vertex.sides}{See ?gplot(sna)}
\item{vertex.rot}{See ?gplot(sna)}
\item{arrowhead.cex}{See ?gplot(sna)}
\item{label.cex}{See ?gplot(sna)}
\item{loop.cex}{See ?gplot(sna)}
\item{vertex.cex}{See ?gplot(sna)}
\item{edge.col}{See ?gplot(sna)}
\item{label.col}{See ?gplot(sna)}
\item{vertex.col}{See ?gplot(sna)}
\item{label.border}{See ?gplot(sna)}
\item{vertex.border}{See ?gplot(sna)}
\item{edge.lty}{See ?gplot(sna)}
\item{label.lty}{See ?gplot(sna)}
\item{vertex.lty}{See ?gplot(sna)}
\item{edge.lwd}{See ?gplot(sna)}
\item{label.lwd}{See ?gplot(sna)}
\item{edge.len}{See ?gplot(sna)}
\item{edge.curve}{See ?gplot(sna)}
\item{edge.steps}{See ?gplot(sna)}
\item{loop.steps}{See ?gplot(sna)}
\item{object.scale}{See ?gplot(sna)}
\item{uselen}{See ?gplot(sna)}
\item{usecurve}{See ?gplot(sna)}
\item{suppress.axes}{See ?gplot(sna)}
\item{vertices.last}{See ?gplot(sna)}
\item{new}{See ?gplot(sna)}
\item{layout.par}{See ?gplot(sna)}
}
\details{See ?gplot(sna).}
\author{Carter T. Butts}